Modularized dyarn internal cmd flag check
